fix(Select): Only dispatch `populated` event when appropriate
This fixes a problem where, when the Select's value is changed (by the
user), the `change` and `populated` events are both dispatched, even if
the list of options is unchanged.

Those subsequent dispatches of the `populated` event make it too easy
for consumers of this component to accidentally set the value of the
Select back to what it was an instant before, effectively preventing
the user's choice from "sticking" sometimes.